# internal/openapi/spec.yaml
openapi: 3.1.0
info:
  title: System Wrangler API
  version: v1
  description: |
    HTTP API exposed by the System Wrangler backend. Every protected
    endpoint requires the `sw_session` cookie issued by
    `POST /api/auth/login` (or `POST /api/auth/setup` on a fresh
    install). State-changing requests must also carry the
    `X-Sw-Csrf: 1` header — the CSRF middleware rejects authenticated
    POST/PUT/PATCH/DELETE without it.

    This spec is the source of truth for the API surface. A drift test
    (`cmd/server/openapi_drift_test.go`) compares it to the routes
    registered on the production mux so the two cannot rot apart.
servers:
  - url: /
    description: Same-origin (the SPA and the API share the host).
...
security:
  - sessionCookie: []
paths:
...
  /api/admin/role-assignments:
    get:
      tags: [rbac]
      summary: List every role assignment. Global Admin only.
      responses:
        "200":
          description: All role assignments, with hydrated username and group name.
          content:
            application/json:
              schema:
                type: object
                required: [assignments]
                properties:
                  assignments:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleAssignment"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [rbac]
      summary: Grant a global or group-scoped role. Global Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAssignmentInput"
      responses:
        "201":
          description: Assignment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignment"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [rbac]
      summary: Revoke a global or group-scoped role. Global Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAssignmentInput"
      responses:
        "204": { description: Assignment revoked. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/groups/{id}/role-assignments:
    get:
      tags: [rbac]
      summary: List role assignments on one group.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Assignments scoped to this group.
          content:
            application/json:
              schema:
                type: object
                required: [assignments]
                properties:
                  assignments:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleAssignment"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [rbac]
      summary: Grant a role on one group. Global Admin or Group Admin of this group.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId: { type: string }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: Assignment created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignment"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/groups/{id}/role-assignments/{userId}/{role}:
    delete:
      tags: [rbac]
      summary: Revoke a role on one group.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: userId, required: true, schema: { type: string } }
        - { in: path, name: role, required: true, schema: { $ref: "#/components/schemas/Role" } }
      responses:
        "204": { description: Assignment revoked. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/me/scope:
    get:
      tags: [rbac]
      summary: Resolved permission set for the caller.
      responses:
        "200":
          description: The caller's global role (if any) and per-group roles.
          content:
            application/json:
              schema:
                type: object
                required: [groups]
                properties:
                  global: { $ref: "#/components/schemas/Role" }
                  groups:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/Role"
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sw_session
  responses:
    BadRequest:
      description: Request payload rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated, but the scope forbids this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist (or is hidden from the caller's scope).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflicts with existing state (duplicate or precondition failed).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Role:
      type: string
      enum: [admin, operator, auditor]
    RoleAssignment:
      type: object
      required: [userId, groupId, role]
      properties:
        userId: { type: string }
        username: { type: string }
        groupId:
          type: [string, "null"]
          description: "`null` for global assignments."
        groupName: { type: string }
        role: { $ref: "#/components/schemas/Role" }
    RoleAssignmentInput:
      type: object
      required: [userId, role]
      properties:
        userId: { type: string }
        groupId:
          type: [string, "null"]
          description: Omit (or `null`) for a global assignment.
        role: { $ref: "#/components/schemas/Role" }
...